[#50292] frontend: third-party: hterm_all: Added an increased timeout for callbacks
During constructor for HTerm there are callbacks for events
such as configuring backgrounds, screen properties etc.

In Firefox, we need to increase the callback timeout limit,
since some of the tasks do not finish (such as setting up screen)
and are silently followed by other tasks that required it (such
as configuring background), which was causing failure of the script.

Signed-off-by: Grzegorz Latosinski <glatosinski@antmicro.com>